Refactored External message processing to allow samples and tests to be processed in the same flow.
Moved the sample and test processing logic into the common base class `AbstractMessageProcessingFlowBase`.
Updated `DoctorDeclarationMessageProcessingFlow` and `LabMessageProcessingFlow` to make use of the shared logic.

Added test type details field to `TestReportDto` to accommodate additional information in case of Other test types.
